backend/app/requirements.txt
pandas==2.2.3
minio==7.2.15
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
boto3==1.37.36
botocore==1.37.36
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
minio==7.2.15
mozjpeg-lossless-optimization==1.1.5
natsort==8.4.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
pycparser==2.22
pycryptodome==3.22.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1
python-slugify==8.0.4
pytz==2025.1
raven==6.10.0
requests==2.32.3
s3transfer==0.11.5
shiboken6==6.8.0.2
six==1.17.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.18
Werkzeug==3.1.3
SQLAlchemy==1.4.23
psycopg2-binary>=2.8,<3.0
Faker
scikit-learn
joblib
scikit-learn
pyyaml
ua-parser
user-agents

# Testing dependencies
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-mock>=3.10.0

# Celery task queue dependencies (Phase 1)
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Backend web server dependencies
gunicorn==21.2.0
eventlet==0.35.2
psycogreen  # Patch psycopg2 for eventlet greenlet compatibility
flask-socketio==5.3.6
python-socketio==5.11.0